ocs_install_verification place change  (#1539)
